move test into internal
