+++ clients/src/test/java/org/apache/kafka/common/utils/internals/ExitTest.java
@@ -14,9 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.common.utils;
-
-import org.apache.kafka.common.utils.internals.Exit;
+package org.apache.kafka.common.utils.internals;
 
 import org.junit.jupiter.api.Test;
